Split session event handlers by category
